JUnit testing implemented

+++ Create-Task-Application/src/main/java/com/batch5/Create_Task_Application/ProjectModule/controller/ProjectController.java
@@ -47,7 +47,7 @@ public class ProjectController {
     // GET /api/projects/user/{userId}
     @GetMapping("/user/{userId}")
     public ResponseEntity<List<ProjectResponseDto>> getProjectsByUser(
-            @PathVariable Integer userId) {
+            @PathVariable Long userId) {
 
         List<ProjectResponseDto> response = projectService.getProjectsByUser(userId);
         return ResponseEntity.ok(response);
